Project filteration and comments added
