--- bin/Main.exe.stackdump
@@ -1,17 +1,0 @@
-Exception: STATUS_ACCESS_VIOLATION at rip=001801B0413
-rax=00000000FFFFCB90 rbx=00000000FFFFCB90 rcx=0000000000000000
-rdx=0000000000000000 rsi=000000010040611A rdi=0000000080008000
-r8 =00000000FFFFCB90 r9 =000000018015F160 r10=0000000100000000
-r11=0000000100401596 r12=000000018025BC20 r13=00000000FFFFCC7C
-r14=0000000000000000 r15=00000000FFFFCC7C
-rbp=00000000FFFFCCD0 rsp=00000000FFFFCB58
-program=C:\Abinav\Proj\bin\Main.exe, pid 6434, thread main
-cs=0033 ds=002B es=002B fs=0053 gs=002B ss=002B
-Stack trace:
-Frame        Function    Args
-000FFFFCCD0  001801B0413 (00000000000, 000FFFFCB1C, 0018015F160, 04949435341)
-000FFFFCCD0  0010040126F (00000000020, FF0700010302FF00, 0018004A7BA, 000FFFFCCD0)
-000FFFFCCD0  0018004A826 (00000000000, 00000000000, 00000000000, 00000000000)
-00000000000  00180048353 (00000000000, 00000000000, 00000000000, 00000000000)
-000FFFFFFF0  00180048404 (00000000000, 00000000000, 00000000000, 00000000000)
-End of stack trace
