+++ doc/login_detail.txt
@@ -1,4 +1,4 @@
-1 hari Admin
+1 vari Admin
 2 abc Student
 3 bb Professor
 4 aa Professor
