+++ include/student.h
@@ -1,1 +1,3 @@
 int student(char*);
+char* raiseTestRequest(char *, int);
+int check_req_status(char *, char *);
